requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
django-multiselectfield==0.1.12
django-resized==0.3.11
Pillow==8.3.1
python-magic-bin==0.4.14
pytz==2021.1
sqlparse==0.4.1
